+++ packages/amazon-cognito-identity-js/index.d.ts
@@ -24,11 +24,20 @@ declare module 'amazon-cognito-identity-js' {
 			userAttributes: any,
 			requiredAttributes: any
 		) => void;
-		mfaRequired?: (challengeName: any, challengeParameters: any) => void;
-		totpRequired?: (challengeName: any, challengeParameters: any) => void;
+		mfaRequired?: (
+			challengeName: ChallengeName,
+			challengeParameters: any
+		) => void;
+		totpRequired?: (
+			challengeName: ChallengeName,
+			challengeParameters: any
+		) => void;
 		customChallenge?: (challengeParameters: any) => void;
-		mfaSetup?: (challengeName: any, challengeParameters: any) => void;
-		selectMFAType?: (challengeName: any, challengeParameters: any) => void;
+		mfaSetup?: (challengeName: ChallengeName, challengeParameters: any) => void;
+		selectMFAType?: (
+			challengeName: ChallengeName,
+			challengeParameters: any
+		) => void;
 	}
 
 	export interface IMfaSettings {
@@ -73,6 +82,14 @@ declare module 'amazon-cognito-identity-js' {
 		CUSTOM_AUTH = 'CUSTOM_AUTH',
 	}
 
+	export type ChallengeName =
+		| 'CUSTOM_CHALLENGE'
+		| 'MFA_SETUP'
+		| 'NEW_PASSWORD_REQUIRED'
+		| 'SELECT_MFA_TYPE'
+		| 'SMS_MFA'
+		| 'SOFTWARE_TOKEN_MFA';
+
 	export interface CognitoUser {
 		attributes: {
 			email: string;
@@ -94,6 +111,8 @@ declare module 'amazon-cognito-identity-js' {
 
 	export class CognitoUser implements CognitoUser {
 		constructor(data: ICognitoUserData);
+
+		challengeName?: ChallengeName;
 
 		public setSignInUserSession(signInUserSession: CognitoUserSession): void;
 		public getSignInUserSession(): CognitoUserSession | null;
@@ -271,8 +290,14 @@ declare module 'amazon-cognito-identity-js' {
 			callbacks: {
 				onSuccess: (session: CognitoUserSession) => void;
 				onFailure: (err: any) => void;
-				mfaRequired?: (challengeName: any, challengeParameters: any) => void;
-				totpRequired?: (challengeName: any, challengeParameters: any) => void;
+				mfaRequired?: (
+					challengeName: ChallengeName,
+					challengeParameters: any
+				) => void;
+				totpRequired?: (
+					challengeName: ChallengeName,
+					challengeParameters: any
+				) => void;
 			}
 		): void;
 	}