+++ packages/pubsub/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.4.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.4.6...@aws-amplify/pubsub@4.4.7) (2022-07-21)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
+## [4.4.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.4.5...@aws-amplify/pubsub@4.4.6) (2022-07-07)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
+## [4.4.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.4.4...@aws-amplify/pubsub@4.4.5) (2022-06-18)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
+## [4.4.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.4.3...@aws-amplify/pubsub@4.4.4) (2022-06-15)
+
+**Note:** Version bump only for package @aws-amplify/pubsub
+
+
+
+
+
 ## [4.4.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pubsub@4.4.2...@aws-amplify/pubsub@4.4.3) (2022-05-24)
 
 **Note:** Version bump only for package @aws-amplify/pubsub